requirements.txt
# AEGIS Core Dependencies
# Phase 1: Data Infrastructure

# Data manipulation
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=12.0.0  # Parquet support

# Exchange API
ccxt>=4.0.0

# Configuration
pyyaml>=6.0

# Retry logic
tenacity>=8.2.0

# Scientific computing
scipy>=1.10.0

# Type hints
typing-extensions>=4.5.0

# Logging
structlog>=23.0.0

# HTTP requests (for webhooks)
requests>=2.31.0

# Time handling
python-dateutil>=2.8.0
pytz>=2023.3

# Development dependencies (optional)
pytest>=7.3.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.3.0
